// src/app/core/models/financeiro/relatorio.model.ts
export interface RelatorioFinanceiro {
  faturamentoPeriodo: number;
  caixaRecebidoPeriodo: number;
  contasAReceber: number;
  contasVencidas: number;
  previsaoRecebimentoProximoMes: number;
  contratosFechadosPeriodo: number;
  ticketMedioPeriodo: number;
  contratosValidos: number;
  contratosPendentes: number;
  dataReferencia: string;
}